requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.40.0
django-appconf==1.0.3
django-compressor==2.3
django-countries==5.5
django-crispy-forms==1.8.0
django-sass-processor==0.7.4
django-tempus-dominus==5.1.2.9
idna==2.8
libsass==0.19.4
oauthlib==3.1.0
Pillow==6.2.1
psycopg2-binary==2.8.4
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.13.0
sqlparse==0.3.0
stripe==2.40.0
urllib3==1.25.7
